requirements.txt
# API
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0
gunicorn==21.2.0

# ML
scikit-learn==1.3.2
pandas==2.1.1
numpy==1.26.0
joblib==1.3.2

# MLflow
mlflow==2.8.0
boto3==1.28.65

# Monitoring
prometheus-client==0.17.1
grafana-api==1.0.3

# Testing
pytest==7.4.3
httpx==0.25.0

# Optimization
onnx==1.15.0
onnxruntime==1.16.0
tensorflow==2.14.0
tensorflow-model-optimization==0.7.5